+++ Changelog.md
@@ -1,7 +1,16 @@
 # version 1
 
-* Recovery time increases multiplier instead of decreasing divider.
-* Recovery time bonus is 10.
-* Attack bonus is 10.
-* Damage bonus is 5.
+* Recovery time bonus is now attack speed bonus.
+* Weapon skill attack per level bonus by rank is 6,8,10.
+* Weapon skill speed per level bonus by rank is 4,6,8.
+* Weapon skill damage per level bonus by rank is 3,4,5.
+* Weapon skill AC per level bonus by rank is 9,12,15.
+* Weapon skill resistance per level bonus by rank is 3,4,5.
+* Mace chance to cause paralysis by rank is 6,8,10.
+* Staff chance to cause slow and feeblemind by rank is 12,16,20.
+* Mace and staff effect duration is 1 minute per skill level.
+* Knight/Cavalier/Champion adds 4/5/6 damage/level to all melee weapon skills.
+* Paladin/Crusader/Hero adds 1/2/3 damage/level to all melee weapon skills.
+* Archer/BattleMage/WarriorMage adds 1/2/3 damage/level to Bow skill.
+* All 4 spells those hit chance was based on skill level now always hit.